Update recommender.py
This is almost final code before the feedback we all have worked on. Since we were having difficulties with the github, we were told by a TA that it would be fine if one person copy pastes the whole thing and add comments on top of the code the person worked on.